+++ components/vf-analytics-google/vf-analytics-google.js
@@ -134,19 +134,21 @@ function vfGaInit() {
  * Track page links as events
  */
 function vfGaLinkTrackingInit() {
-  // jQuery("body.google-analytics-loaded .track-with-analytics-events a").on('mousedown', function(e) {
-  //   analyticsTrackInteraction(e.target,'Manually tracked area');
-  // });
-
-  document.body.onclick = function(e){
-    e = e || event;
-    var from = findParent('a',e.target || e.srcElement);
-    if (from){
+  document.body.addEventListener("mousedown", function (evt) {
+    // send GA events if GA closest area is detected
+    let closestContainer = getClosestGa(evt.target, '[data-vf-google-analytics-region]');
+    if (closestContainer) {
+      vfGaTrackInteraction(evt.target);
+    } else {
+      var from = findParent('a',evt.target || evt.srcElement);
+      if (from) {
        /* it's a link, actions here */
-       console.log('click',from);
-       analyticsTrackInteraction(from);
+       // console.log('clicked from findParent: ',from);
+       vfGaTrackInteraction(from);
+      }
     }
-  }
+  }, false );
+
   //find first parent with tagName [tagname]
   function findParent(tagname,el){
     while (el){
@@ -159,6 +161,24 @@ function vfGaLinkTrackingInit() {
   }
 }
 
+/*
+ * Find closest element that has GA attribute
+ * @returns {el} the closest element with GA attribute
+ */
+function getClosestGa(elem, selector) {
+  // Element.matches() polyfill
+  // https://developer.mozilla.org/en-US/docs/Web/API/Element/matches
+  if (!Element.prototype.matches) {
+    Element.prototype.matches = Element.prototype.msMatchesSelector || Element.prototype.webkitMatchesSelector;
+  }
+
+	// Get the closest matching element
+	for ( ; elem && elem !== document; elem = elem.parentNode ) {
+		if ( elem.matches( selector ) ) return elem;
+  }
+	return null;
+};
+
 /**
  * Utility method to get the last in an array
  * @returns {var} the last item in the array
@@ -170,11 +190,17 @@ if (!Array.prototype.vfGaLinkLast){
   };
 };
 
+// Catch any use cases that may have been existing
+// To be removed in 2.0.0
+function analyticsTrackInteraction(actedOnItem, customEventName) {
+  console.warn('vfGa','As of 1.0.0-rc.3 analyticsTrackInteraction() is now vfGaTrackInteraction(). You function call is being proxied. You should update your code.');
+  vfGaTrackInteraction(actedOnItem, customEventName);
+}
+
 /**
  * Analytics tracking
  * ---
- * This code tracks the user's clicks in various parts of the site and logs them as GA events.<br/>
- * Links in non-generic regions can be tracked by adding '.track-with-analytics-events' to a parent div. Careful with the scoping.
+ * This code tracks the user's clicks in various parts of the site and logs them as GA events.
  *
  * Dev note:
  * add class verbose-analytics to your body for a readout to console on clicks.
@@ -183,22 +209,40 @@ if (!Array.prototype.vfGaLinkLast){
  * @param {string} customEventName Event action
  * @example
  * jQuery(".analytics-content-footer").on('mousedown', 'a, button', function(e) {
- *   analyticsTrackInteraction(e.target,'Content footer');
+ *   vfGaTrackInteraction(e.target,'Content footer');
  * });
  */
-function analyticsTrackInteraction(actedOnItem, customEventName) {
+function vfGaTrackInteraction(actedOnItem, customEventName) {
   var customEventName = customEventName || []; // you can pass some custom text as a 3rd param
   let linkName;
 
   if (customEventName.length > 0) {
     linkName = customEventName;
   } else { // then derive a value
+
+    // Fix for when tags have undefined .innerText
+    if (typeof actedOnItem.innerText === 'undefined') {
+      actedOnItem.innerText = '';
+    }
+
     linkName = actedOnItem.innerText;
-    console.log('linkName',linkName);
+    // console.log('linkName',linkName);
 
     // if there's no text, it's probably and image
     if (linkName.length == 0 && actedOnItem.hasAttribute('src')) linkName = actedOnItem.src.split('/').vfGaLinkLast();
     if (linkName.length == 0 && actedOnItem.value) linkName = actedOnItem.value;
+
+    // is there an inner image?
+    if (linkName.length == 0 && actedOnItem.getElementsByTagName('img')) {
+      if (actedOnItem.getElementsByTagName('img')[0]) {
+        if (actedOnItem.getElementsByTagName('img')[0].hasAttribute('src')) {
+          linkName = actedOnItem.src.split('/').vfGaLinkLast();
+        }
+      }
+    }
+
+    // fallback to an href value
+    if (linkName.length == 0 && actedOnItem.href) linkName = actedOnItem.href;
 
     // special things for gloabl search box
     // if (parentContainer == 'Global search') {
@@ -208,10 +252,10 @@ function analyticsTrackInteraction(actedOnItem, customEventName) {
 
   // Get closest parent container
   // Track the region of the link clicked (global nav, masthead, hero, main content, footer, etc)
-  //data-vf-google-anlaytics-region="main-content-area-OR-SOME-OTHER-NAME"
-  let parentContainer = actedOnItem.closest("[data-vf-google-anlaytics-region]");
+  //data-vf-google-analytics-region="main-content-area-OR-SOME-OTHER-NAME"
+  let parentContainer = actedOnItem.closest("[data-vf-google-analytics-region]");
   if (parentContainer) {
-    parentContainer = parentContainer.dataset.vfGoogleAnlayticsRegion;
+    parentContainer = parentContainer.dataset.vfGoogleAnalyticsRegion;
   } else {
     parentContainer = 'No container specified';
   }
@@ -225,30 +269,49 @@ function analyticsTrackInteraction(actedOnItem, customEventName) {
 
     // Track file type (PDF, DOC, etc) or if mailto
     // adapted from https://www.blastanalytics.com/blog/how-to-track-downloads-in-google-analytics
-    var filetypes = /\.(zip|exe|pdf|doc*|xls*|ppt*|mp3)$/i;
+    var filetypes = /\.(zip|exe|pdf|doc*|xls*|ppt*|mp3|txt|fasta)$/i;
     var baseHref = '';
     var href = actedOnItem.href;
     if (href && href.match(/^mailto\:/i)) {
       var mailLink = href.replace(/^mailto\:/i, '');
       ga && ga('send', 'event', 'Email', 'Region / ' + parentContainer, mailLink);
+      // Log email event
+      vfGaLogMessage("Email", "Region / " + parentContainer, mailLink, lastGaEventTime, actedOnItem);
     }
     if (href && href.match(filetypes)) {
       var extension = (/[.]/.exec(href)) ? /[^.]+$/.exec(href) : undefined;
       var filePath = href;
       ga && ga('send', 'event', 'Download', 'Type / ' + extension + ' / ' + parentContainer, filePath);
+      // Log Download event
+      vfGaLogMessage("Download", "Type / " + extension + " / " + parentContainer, filePath, lastGaEventTime, actedOnItem);
     }
 
     // note that we've stored an event(s)
     lastGaEventTime = Date.now();
+    // Default log the event.
+    vfGaLogMessage("UI", "UI Element / " + parentContainer, linkName, lastGaEventTime, actedOnItem);
+  }
+}
 
-    // conditional logging
-    let conditionalLoggingCheck = document.querySelector('body');
-    // debug: always turn on verbos analytics
-    // conditionalLoggingCheck.setAttribute('data-vf-google-analytics-verbose', 'true');
-    if (conditionalLoggingCheck.dataset.vfGoogleAnalyticsVerbose) {
-      console.log('%c Verbose analytics on ', 'color: #FFF; background: #111; font-size: .75rem;');
-      console.log('clicked on: %o ',actedOnItem);
-      console.log('sent to GA: ', 'event ->', 'UI ->', 'UI Element / ' + parentContainer + ' ->', linkName, '; at: ',lastGaEventTime);
+/**
+ * Helper function to log debug console messages.
+ *
+ * @param {string} eventCategory
+ * @param {string} eventAction
+ * @param {string} eventLabel
+ * @param {string} lastGaEventTime
+ * @param {element} actedOnItem
+ */
+function vfGaLogMessage(eventCategory, eventAction, eventLabel, lastGaEventTime, actedOnItem) {
+  // conditional logging
+  let conditionalLoggingCheck = document.querySelector("body");
+  // debug: always turn on verbose analytics
+  // conditionalLoggingCheck.setAttribute('data-vf-google-analytics-verbose', 'true');
+  if (conditionalLoggingCheck.dataset.vfGoogleAnalyticsVerbose) {
+    if (conditionalLoggingCheck.dataset.vfGoogleAnalyticsVerbose == "true") {
+      console.log("%c Verbose analytics on ", "color: #FFF; background: #111; font-size: .75rem;");
+      console.log("clicked on: %o ", actedOnItem);
+      console.log("sent to GA: ", "event ->", eventCategory + " ->", eventAction + " ->", eventLabel, "; at: ", lastGaEventTime);
     }
   }
 }