+++ tools/vf-component-library/eleventy.js
@@ -1,0 +1,91 @@
+const { DateTime } = require('luxon');
+const _            = require('lodash');
+const Path         = require('path');
+
+module.exports = function(config) {
+
+  // Add in tags, filters useful for Visual Framework installs
+  // (fractal's render tag, codeblock, markdown, etc)
+  // and common configuration
+  const vfEleventyExtension = require("@visual-framework/vf-extensions\/11ty");
+  config.addPlugin(vfEleventyExtension);
+
+  // BroswerSync options
+  config.setBrowserSyncConfig({ open: true });
+
+  // Filters
+  // https://www.11ty.io/docs/filters/
+  // -----
+
+  // {{ "myContent" | sampleFilter }}
+  // config.addFilter("sampleFilter", function(value) {
+  //   return 'ddd' + value;
+  // });
+
+  // Add any utiliuty filters
+  config.addFilter("dateDisplay", (dateObj, format = "d LLL y") => {
+    return DateTime.fromJSDate(dateObj, {
+      zone: "utc"
+    }).toFormat(format);
+  });
+
+  // Shortcodes
+  // https://www.11ty.io/docs/shortcodes/
+  // -----
+
+  // nunjucks
+  // {% sampleShortcode "firstName", "lastName" %}
+  // handlebars
+  // {{ sampleShortcode "firstName" "lastName" }}
+  // config.addShortcode("sampleShortcode", function(firstName, lastName) {
+  //   return 'hi ' + firstName + lastName;
+  // });
+
+  // If you want to minify html output
+  // config.addTransform("htmlmin", require("./node_modules/\@visual-framework/vf-eleventy--extensions/utils/minify-html.js"));
+
+  // Add any custom tags
+
+  // config.addNunjucksTag("uppercase", function(nunjucksEngine) {
+  //   return new function() {
+  //     this.tags = ["uppercase"];
+  //
+  //     this.parse = function(parser, nodes, lexer) {
+  //       var tok = parser.nextToken();
+  //
+  //       var args = parser.parseSignature(null, true);
+  //       parser.advanceAfterBlockEnd(tok.value);
+  //
+  //       return new nodes.CallExtensionAsync(this, "run", args);
+  //     };
+  //
+  //     this.run = function(context, myStringArg, callback) {
+  //       let ret = new nunjucksEngine.runtime.SafeString(
+  //         myStringArg.toUpperCase()
+  //       );
+  //       callback(null, ret);
+  //     };
+  //   }();
+  // });
+
+  // pass some assets right through
+  // config.addPassthroughCopy("./src/site/images");
+
+  return {
+    dir: {
+      input: "src/site",
+      output: "build",
+      data: "_data"
+    },
+    templateFormats : [
+      "njk", "md", // note that .md files will also be parsed with njk processor
+      "css", "js" // passthrough file copying for static assets
+    ],
+    openBrowser: true,
+    htmlTemplateEngine : ["njk", "md"],
+    markdownTemplateEngine : "njk",
+    passthroughFileCopy: true,
+    pathPrefix: "/vf-core/" // if your site is deployed to a sub-url, otherwise comment out
+  };
+
+};