+++ tools/vf-component-library/src/site/_data/serverInfo.js
@@ -1,0 +1,4 @@
+// https://www.11ty.io/docs/data-js/
+module.exports = {
+  environment: process.env.ELEVENTY_ENV
+};