+++ components/ebi-header-footer/CHANGELOG.md
@@ -1,5 +1,11 @@
-# Change Log
+### 1.0.1
 
-## 1.0.0-beta.2
+- adds `webkit-appearance: none;` as needed for Safari browsers as autoprefixer is not doing this.
 
-* Resolves changes in other VF components, spacing, links, etc. 
+### 1.0.0
+
+- EBI Header is once again fullbleed after changes in vf-grid-page 2.0.0
+
+### 1.0.0-beta.2
+
+- Resolves changes in other VF components, spacing, links, etc.