+++ components/vf-form/vf-form__legend/CHANGELOG.md
@@ -1,5 +1,3 @@
-# Change Log
-
-## 1.0.0
+### 1.0.0
 
 - initial commit of new legend component, that acts like `vf-form__label`.