+++ components/vf-intro/CHANGELOG.md
@@ -1,9 +1,32 @@
-# Change Log
+### 1.3.2
 
-## 1.0.2
+- resets the custom property used for `gap` in `embl-grid` so it only worries about the column gap.
 
-* Simplify grid
+### 1.3.1
 
-## 1.0.0 (2019-12-17)
+- updates max-width of component
 
-* Initial stable release
+### 1.3.0
+
+- adds ability for `vf-badge` to be customised with it's variants
+- allows for fallbacks so older versions will still work
+- removes old way of doing `vf-badge` from demo code
+- makes `vf-lede` work.
+
+### 1.2.0
+
+- wraps all content except heading in if statement
+- changes how badges are implemented inside an if statement
+
+### 1.1.0
+
+- adds if statement to pass context data through to component
+- improves documentation
+
+### 1.0.2
+
+- Simplify grid
+
+### 1.0.0
+
+- Initial stable release