+++ components/vf-sass-starter/CHANGELOG.md
@@ -1,10 +1,24 @@
-# Change Log
+### 0.1.7
 
-## 0.1.0
+- dependency bump
 
-* updates package.json to dependencies rather than devDependencies
-* updates relative paths of index.scss to target installed packages in node_modules
+### 0.1.4
 
-## 0.0.3
+- dependency bump
 
-Initial version
+### 0.1.2
+
+- dependency bump
+
+### 0.1.1
+
+- dependency bump
+
+### 0.1.0
+
+- updates package.json to dependencies rather than devDependencies
+- updates relative paths of index.scss to target installed packages in node_modules
+
+### 0.0.3
+
+- Initial version